+++ src/main/java/dk/mfaester/grav/Box.java
@@ -25,8 +25,8 @@ public class Box extends AbstractDrawable {
     }
 
     @Override
-    protected byte[] createIndices() {
-        final byte[] indices = {
+    protected int[] createIndices() {
+        final int[] indices = {
                 0, 1, 2,
                 3, 2, 0
         };
